chore: remove unused files, rename files and create create chat route
