+++ src/functions/websocket/handlers/message.ts
@@ -6,6 +6,7 @@ import { chats } from "../../../drizzle/schema/chats";
 import { messages } from "../../../drizzle/schema/messages";
 import { z } from "zod";
 import { app } from "../../../server";
+import { wss } from "../../../common/websocket";
 
 const messageSchema = z.object({
 	chatId: z.number(),
