+++ src/routes/auth/new-user-route.ts
