+++ src/server.ts
@@ -8,9 +8,9 @@ import {
 	serializerCompiler,
 } from "fastify-type-provider-zod";
 import { env } from "./common/env";
-import { createUserRoute } from "./routes/auth/new";
-import { sendCodeToUserRoute } from "./routes/auth/send";
-import { sendMessageRoute } from "./routes/messages/send-message";
+import { createUserRoute } from "./routes/auth/new.-user-route";
+import { sendCodeToUserRoute } from "./routes/auth/send-code-route";
+import { createChatRoute } from "./routes/chats/create-chat-route";
 
 export const app = fastify().withTypeProvider<ZodTypeProvider>();
 app.register(fastifyCors, { credentials: true }); //pra enviar cookie pro front
@@ -29,12 +29,12 @@ app.setValidatorCompiler(validatorCompiler);
 //validador e serializador do zod, pro fastify conseguir ler e enviar os dados usando o zod
 app.register(createUserRoute);
 app.register(sendCodeToUserRoute);
-app.register(sendMessageRoute);
+app.register(createChatRoute);
 //escolhe a porta que vai ser aberta pra API e abre essa porta (e o console.log pra avisar que subiu)
 app
 	.listen({ port: env.PORT })
 	.then(() => {
-		console.log("HTTP Server running!");
+		console.log("HTTP/Websocket Server running!");
 	})
 	.catch((e) => {
 		console.log(e);
